% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balanced.folds.R
\name{balanced.folds}
\alias{balanced.folds}
\title{Balanced cross-validation segments}
\usage{
balanced.folds(y, nfolds = min(min(table(y)), 10))
}
\arguments{
\item{y}{A categorical vector of class memberships}

\item{nfolds}{The required number of CV-segements or "folds".}
}
\value{
A list of sample segments.
}
\description{
Internal function for constructing a class-balanced segment list for
cross-validation (copy from a previous verson from package pamr).
}
\examples{

data(BCdata)
balanced.folds(BCdata$Y)
}
\keyword{cross-validation}
\keyword{folds}
