% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\docType{data}
\name{BCdata}
\alias{BCdata}
\title{BCdata Documentation
BCdata is the data used in this package}
\usage{
data(BCdata)
}
\description{
BCdata Documentation
BCdata is the data used in this package
}
\examples{
data(BCdata)
}
\keyword{BCdata}
\keyword{datasets}
\keyword{lpls-data}
