% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centering.R
\name{centering}
\alias{centering}
\title{Matrix centering}
\usage{
centering(Mat, rowcenter = FALSE, colcenter = FALSE, grandcenter = FALSE)
}
\arguments{
\item{Mat}{A numerical matrix.}

\item{rowcenter}{Logical. If \code{TRUE} the matrix is row-centered}

\item{colcenter}{Logical. If \code{TRUE} the matrix is columns centered}

\item{grandcenter}{Logical. If \code{TRUE} the matrix is only centered by a
subtraction by the overall (grand) mean of all matrix elements. Only applies
if both rowcenter and colcenter are \code{FALSE}.}
}
\value{
A centered matrix
}
\description{
A function which performs row-,  column-,  double- or overall-centering of a
matrix
}
\details{
If both row- and colcenter are \code{TRUE} the matrix is double centered.
This is done by first subtracting the pre-computed row-means,  then the
pre-computed column means and finally,  the grand mean is added back in.
}
\examples{

data(BCdata)
Xc <- centering(BCdata$X,  FALSE,  TRUE,  FALSE)
}
\author{
Solve S<c3><a6>b<c3><b8>
}
\keyword{centering}
\keyword{lpls}
