% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confusion.R
\name{confusion}
\alias{confusion}
\title{Error rate and confusion matrix}
\usage{
confusion(trueclass, predclass, labs)
}
\arguments{
\item{trueclass}{The true classes}

\item{predclass}{The predicted classes}

\item{labs}{A vector of labels for the classes}
}
\value{
\item{total.error }{The total classification error}
\item{conf.matrix }{The (g x g) confusion matrix}
}
\description{
The total error and the (g x g) confusion matrix for g-class classification
problems
}
\examples{

data(BCdata)
confusion(factor(BCdata$Y), sample(c(0, 1), 130, replace = TRUE), c("0", "1"))
}
\author{
Solve Saebo
}
\keyword{classification}
\keyword{error}
