% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractscores.R
\name{extractscores}
\alias{extractscores}
\title{Latent components extraction via NIPALS.}
\usage{
extractscores(X1, X2, X3, niter = 5, threshold = 0)
}
\arguments{
\item{X1}{A data matrix (n x q) whose rows is connected to the rows of X2}

\item{X2}{A data matrix (n x p) whose rows is connected to the rows of X1,
and whose columns are connected to the columns of X3.}

\item{X3}{A data matrix (m x p) whose columns are connected toe the columns
of X2.}

\item{niter}{Number of NIPALS iterations.}

\item{threshold}{A threshold parameter in [0, 1) for potential
soft - thresholding of the rows of X3.}
}
\value{
Returns a list og six latent vectors (normalized).
}
\description{
Function for extracting latent components from L - structure data using the
NIPALS algorithm.
}
\examples{

data(BCdata)
scores <- extractscores(BCdata$Y,  BCdata$X,  BCdata$Z)
str(scores)
}
\author{
Solve Sæbø
}
\keyword{NIPALS}
\keyword{latent}
\keyword{lpls}
