% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm.R
\name{norm}
\alias{norm}
\title{Vector normalization}
\usage{
norm(vec)
}
\arguments{
\item{vec}{A numerical vector}
}
\value{
The normalized vector
}
\description{
Normalizes a vector to have length equal to 1.
}
\examples{

data(BCdata)
x <- norm(BCdata$X[,1])
crossprod(x)
}
\author{
Solve Sæbø
}
\keyword{lpls}
\keyword{normalization}
