% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.rows.R
\name{permute.rows}
\alias{permute.rows}
\title{Internal function (copied from previous version of the package pamr)}
\usage{
permute.rows(x)
}
\arguments{
\item{x}{A numerical matrix}
}
\value{
A matrix with permuted columns within rows.
}
\description{
Function for column permutation within rows of a matrix
}
\examples{

data(BCdata)
permute.rows(BCdata$X[1:10,1:5])
}
\keyword{lpls}
\keyword{permutation}
