% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.lplsReg.dcv.R
\name{plot.lplsReg.dcv}
\alias{plot.lplsReg.dcv}
\title{Double-CV plot}
\usage{
\method{plot}{lplsReg.dcv}(object, identifyVariable = FALSE)
}
\arguments{
\item{object}{A double-Cv object as returned from \code{lplsReg.dcv}}

\item{identifyVariable}{Logical. Should interactive variable identification
be activated?}
}
\description{
Plot function summarizing some results from double-crossvalidation of
lpls-regression/classification.
}
\details{
The first plot is a plot of posterior probability of class membership for
each sample plotted versus sample number. For g-group classification, there
will be g dots per sample, and the largest dot indicates the predicted
class/group. The second plot is a plot of the frequency of each variable
being selected by jack-knifing in each of the cross-validation segments.
High selection frequency may be considered as a measure of variable
importance and stability.
}
\author{
Solve S<c3><a6>b<c3><b8>
}
\keyword{posterior}
\keyword{variable-selection}
