% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.lplsReg.R
\name{predict.lplsReg}
\alias{predict.lplsReg}
\title{Function for lpls prediction.}
\usage{
\method{predict}{lplsReg}(object, ncomp = object$npc, X2new = NULL,
  silent = FALSE)
}
\arguments{
\item{object}{A fitted lpls-model as returned from \code{lplsReg}}

\item{ncomp}{The number of components to be used for prediction.}

\item{X2new}{A matrix of size (l x p) holding the predictor values of X2 for
l new observations.}

\item{silent}{Suppress warning about components being large}
}
\value{
\item{pred}{The predicted response values.} 
\item{beta0 }{The estimated intercept in the lpls-regression model} 
\item{beta1}{The estimated vector of regression coefficients for the p 
predictor variables in X2.}
}
\description{
The function predicts new responses based on a fitted lpls-model and a set
of new predictor data.
}
\examples{

data(BCdata)
fit  <- lplsReg(BCdata$Y, BCdata$X, BCdata$Z, npc = 10)
pred <- predict(fit, 5, X2new = BCdata$X[1:10,])
pred$pred
}
\author{
Solve Saebo
}
\keyword{lpls}
\keyword{prediction}
