% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectonto.R
\name{projectonto}
\alias{projectonto}
\title{Internal function for projections}
\usage{
projectonto(A, b)
}
\arguments{
\item{A}{A centered matrix whose columns will be projected onto}

\item{b}{this centered vector}
}
\value{
A vector of slope coefficients.
}
\description{
Projects the (centred) columns of a matrix onto a (centered) vector and
returns the slope-coefficient for each column.
}
\examples{


data(BCdata)
Y <- BCdata$Y
X <- BCdata$X
#Project the columns of X onto Y.
tt <- projectonto(X,Y)
}
\author{
Solve S<c3><a6>b<c3><b8>
}
\keyword{lpls}
\keyword{projection}
