% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lplsReg.dcv.R
\name{summary.lplsReg.dcv}
\alias{summary.lplsReg.dcv}
\title{Summary function for lplsReg double cross-validation objects.}
\usage{
\method{summary}{lplsReg.dcv}(obj)
}
\arguments{
\item{obj}{An object returned from a call to \code{lplsReg.dcv}.}
}
\value{
Prints the call, the total classification error and the confusion
matrix.
}
\description{
Summary function for lplsReg double cross-validation objects.
}
\author{
Solve Sæbø
}
\keyword{classification}
\keyword{error-rate}
